old-tests/submission/W3C/domhtml/HTMLAppletElement05.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - Applet</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-applet-codebase" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<applet align="bottom" alt="Applet Number 1" archive="" code="org/w3c/domts/DOMTSApplet.class" codebase="applets" height="306" hspace="0" name="applet1" vspace="0" width="301"></applet>
</p>


<script>
var domtest = async_test("The codeBase IDL attribute must reflect the codebase content attribute, which for the purposes of reflection is defined as containing a URL.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vcodebase;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("applet");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vcodebase = testNode.codeBase;

      assert_uri_equals(document.location.protocol,null,document.location.host,"applets",null,null,null,true,vcodebase);

       
    domtest.done();
}));

</script>
<p>The codeBase IDL attribute must reflect the codebase content attribute, which for the purposes of reflection is defined as containing a URL.</p>
<div id="log"></div>
</body></html>
